requirements.txt
# Core Django
Django==5.1.3
psycopg2-binary==2.9.10

# Django Extensions
django-extensions==3.2.3
django-environ==0.11.2
environs[django]==11.0.0

# REST Framework & API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
drf-spectacular==0.27.2

# Authentication & Social Auth
django-allauth==65.1.0
PyJWT[crypto]==2.9.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.156.0

# File & Media Management
Pillow==11.0.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-storages==1.14.4

# PDF Generation
weasyprint==52.5

# Office Document Handling
python-docx==1.1.2
XlsxWriter==3.2.0

# HTTP & Networking
requests==2.31.0
httpx==0.27.0

# Caching & Performance
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2

# Background Tasks
django-background-tasks==1.2.8

# API Utilities
django-filter==24.3
django-cors-headers==4.3.1
django-ratelimit==4.1.0

# Caching Extensions
drf-extensions==0.8.0

# Static Files
whitenoise==6.8.2

# Server
gunicorn==23.0.0

# Development Tools
django-debug-toolbar==4.4.6
pyngrok==7.5.0

# Utilities
pytz==2024.2
sqlparse==0.5.1
asgiref==3.8.1
typing_extensions==4.12.2

# Security
cryptography==42.0.5